requirements.txt
django>=4.0,<5.0
psycopg2-binary>=2.9.3
google-api-python-client>=2.50.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-auth>=2.6.0
django-allauth>=0.54.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
python-dotenv>=0.20.0
PyJWT>=2.10.1
cryptography>=41.0.0
requests>=2.29.0
croniter>=1.4.0
openai>=1.1.0
django-cryptography>=1.1.0
djangorestframework==3.14.0
markdown==3.5.1       # For browsable API docs
django-filter==23.3   # For filtering support
coreapi==2.3.3        # For API documentation

# django-debug-toolbar>=5.2.0
